Automated attendance system using face recognition and MYSQL database
Nowadays Educational institutions are concerned about regularity of student attendance. Even in pandemic situation attendance is still a major issue in schools and colleges. Mainly there are two conventional methods of marking attendance which are calling out the roll call or by taking student sign on paper. They both were more time consuming and difficult. Hence, there is a requirement of computer-based student attendance management system which will assist the faculty for maintaining attendance record automatically. In this project we have implemented the automated attendance system using 'PYTHON' and 'MYSQL'. We have projected our ideas to implement an “Automated Attendance System Based on Face Recognition”. The application includes face identification, which saves time as well as being purely software based it can be flagged as eco-friendly as it reduces the use of paper. This system also eliminates the chances of fake attendance because of the face being used as a biometric for authentication. Hence, this system can be implemented in a field where attendance plays an important role. The proposed system is designed using PYTHON as well as SQL database. The algorithm used in the system is based on image comparison on the basis of the encoded values of the face from the image from database with the image recorded by the system in run time. The system has output in the form of MYSQL DATABASE.